update to fex github.com/kode54/File_Extractor/commit/e9e7560df9ca9b25348e0bfb4012f6e570711bb3 should fix new format RARs that dont work, but havent tested it

+++ trunk/desmume/src/windows/File_Extractor/7z_C/7zIn.c
@@ -1322,7 +1322,7 @@ SRes SzArEx_Open(CSzArEx *p, ILookInStream *inStream, ISzAlloc *allocMain, ISzAl
 SRes SzArEx_Extract(
     const CSzArEx *p,
     ILookInStream *inStream,
-    UInt32 fileIndex,
+    UInt64 fileIndex,
     UInt32 *blockIndex,
     Byte **outBuffer,
     size_t *outBufferSize,
